+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `assertJsonPathEmpty()`, `assertJsonPathNotEmpty()`,
+  `assertJsonPathContains()`, and `assertJsonPathNotContains()` methods to the
+  test response class.
+
 ### Changed
 
 - Changed the `theme()` method of the installation manager to set the